requirements.txt
openai
python-dotenv
PyAudio
streamlit
SpeechRecognition
opencv-python
opencv-contrib-python
fastapi
twilio